:basketball:

+++ leetcode-new/SuperPow.java
@@ -20,7 +20,7 @@ public class SuperPow {
         int result = 0;
         for (int i = 0, base = a; i < b.length; ++i) {
             result += b[b.length - 1 - i] * base;
-            base = Math.pow(base, 10);
+            base = (int) Math.pow(base, 10);
         }
         return result;
     }
